feat(splineROI): moved all types to tools/types and updated example-info
