+++ packages/tools/src/types/CardinalSplineProps.ts
