+++ packages/tools/src/types/ClosestControlPoint.ts
