+++ packages/tools/src/types/ClosestPoint.ts
