+++ packages/tools/src/types/ClosestSplinePoint.ts
