+++ packages/tools/src/types/SplineCurveSegment.ts
