+++ packages/tools/src/types/SplineLineSegment.ts
