+++ packages/tools/src/types/SplineProps.ts
